# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hyderabad Urban Flood Nowcasting System foundation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /safe-route:
    get:
      operationId: getSafeRoute
      tags: [nowcasting]
      summary: Compare normal and flood-aware routes
      parameters:
        - {name: source_lat, in: query, required: true, schema: {type: number}}
        - {name: source_lon, in: query, required: true, schema: {type: number}}
        - {name: destination_lat, in: query, required: true, schema: {type: number}}
        - {name: destination_lon, in: query, required: true, schema: {type: number}}
        - {name: forecast_minutes, in: query, required: false, schema: {type: integer, enum: [0, 30, 60, 90, 120, 150, 180], default: 0}}
      responses:
        "200":
          description: Normal and flood-aware route comparison
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SafeRouteResponse"
        "400":
          description: Invalid coordinates or unavailable route
  /map-layers:
    get:
      operationId: getMapLayers
      tags: [nowcasting]
      summary: Get simplified web map layers
      description: Returns cached lightweight GeoJSON layers and downsampled forecast cells; scientific rasters and the full road graph are not sent to the browser.
      parameters:
        - {name: forecast_minutes, in: query, required: false, schema: {type: integer, enum: [0, 30, 60, 90, 120, 150, 180], default: 0}}
      responses:
        "200":
          description: Simplified map layers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MapLayersResponse"
...
  /runoff-summary:
    get:
      operationId: getRunoffSummary
      tags: [nowcasting]
      summary: Get rainfall-to-runoff summary statistics
      description: Returns statistics only after a successful runoff model run.
      responses:
        "200":
          description: Runoff summary statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunoffSummaryResponse"
        "404":
          description: Runoff model has not successfully run
...
  /surface-water-summary:
    get:
      operationId: getSurfaceWaterSummary
      tags: [nowcasting]
      summary: Get prototype surface-water summary statistics
      description: Returns real statistics from the generated terrain-based surface-water rasters.
      responses:
        "200":
          description: Surface-water summary statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SurfaceWaterSummaryResponse"
        "404":
          description: Surface-water model has not successfully run
...
  /drainage-summary:
    get:
      operationId: getDrainageSummary
      tags: [nowcasting]
      summary: Get prototype drainage interaction statistics
      responses:
        "200":
          description: Drainage model summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrainageSummaryResponse"
        "404":
          description: Drainage model has not successfully run
...
  /flood-depth-summary:
    get:
      operationId: getFloodDepthSummary
      tags: [nowcasting]
      summary: Get prototype flood-depth screening statistics
      responses:
        "200":
          description: Flood-depth summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FloodDepthSummaryResponse"
        "404":
          description: Flood-depth model has not successfully run
  /road-flood-risk:
    get:
      operationId: getRoadFloodRisk
      tags: [nowcasting]
      summary: Get processed road flood-risk artifact status
      description: Returns metadata and the path for the generated road_flood_risk GeoJSON artifact.
      responses:
        "200":
          description: Road flood-risk summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadFloodRiskResponse"
        "404":
          description: Flood-depth model has not successfully run
components:
  schemas:
...
    RoutePoint:
      type: object
      properties:
        lat: {type: number}
        lon: {type: number}
      required: [lat, lon]
    RouteSummary:
      type: object
      properties:
        node_sequence:
          type: array
          items: {type: string}
        route:
          type: array
          items: {$ref: "#/components/schemas/RoutePoint"}
        distance_m: {type: number}
        estimated_time_min: {type: number}
        max_flood_depth_cm: {type: number}
        mean_flood_depth_cm: {type: number}
        max_risk_class: {type: integer}
        safety: {type: string}
      required: [node_sequence, route, distance_m, estimated_time_min, max_flood_depth_cm, mean_flood_depth_cm, max_risk_class, safety]
    SafeRouteResponse:
      type: object
      properties:
        status: {type: string}
        forecast_minutes: {type: integer}
        distance_m: {type: number}
        estimated_time_min: {type: number}
        max_flood_depth_cm: {type: number}
        mean_flood_depth_cm: {type: number}
        max_risk_class: {type: integer}
        safety: {type: string}
        flood_avoided: {type: boolean}
        route:
          type: array
          items: {$ref: "#/components/schemas/RoutePoint"}
        normal_route: {$ref: "#/components/schemas/RouteSummary"}
        flood_safe_route: {$ref: "#/components/schemas/RouteSummary"}
        warnings:
          type: array
          items: {type: string}
      required: [status, forecast_minutes, route, normal_route, flood_safe_route, safety]
    GeoJsonFeatureCollection:
      type: object
      additionalProperties: true
      properties:
        type: {type: string}
        features:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [type, features]
    MapLayersResponse:
      type: object
      properties:
        status: {type: string}
        forecast_minutes: {type: integer}
        layers:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/GeoJsonFeatureCollection"
        limits:
          type: object
          additionalProperties: true
      required: [status, forecast_minutes, layers, limits]
...
    RunoffSummaryResponse:
      type: object
      properties:
        status:
          type: string
        minimum_runoff_depth_mm:
          type: number
          nullable: true
        maximum_runoff_depth_mm:
          type: number
          nullable: true
        mean_runoff_depth_mm:
          type: number
          nullable: true
        total_runoff_volume_m3:
          type: number
          nullable: true
        valid_cell_count:
          type: integer
        rainfall_provider:
          type: string
        rainfall_timestamp:
          type: string
        rainfall_scenario:
          type: string
        crs:
          type: string
      required:
        - status
        - minimum_runoff_depth_mm
        - maximum_runoff_depth_mm
        - mean_runoff_depth_mm
        - total_runoff_volume_m3
        - valid_cell_count
        - rainfall_provider
        - rainfall_timestamp
        - rainfall_scenario
        - crs
...
    SurfaceWaterSummaryResponse:
      type: object
      properties:
        status:
          type: string
        rainfall_timestamp:
          type: string
        rainfall_scenario:
          type: string
        valid_cell_count:
          type: integer
        maximum_surface_water_depth_mm:
          type: number
          nullable: true
        maximum_surface_water_depth_cm:
          type: number
          nullable: true
        mean_surface_water_depth_mm:
          type: number
          nullable: true
        total_surface_water_volume_m3:
          type: number
        threshold_cell_counts:
          type: object
          additionalProperties:
            type: integer
        nala_interaction_cell_count:
          type: integer
        water_body_cell_count:
          type: integer
        input_runoff_volume_m3:
          type: number
        prototype_drainage_removed_volume_m3:
          type: number
        volume_conservation_error_m3:
          type: number
        volume_conservation_check:
          type: boolean
      required:
        - status
        - valid_cell_count
        - maximum_surface_water_depth_mm
        - maximum_surface_water_depth_cm
        - mean_surface_water_depth_mm
        - total_surface_water_volume_m3
        - threshold_cell_counts
        - nala_interaction_cell_count
        - water_body_cell_count
        - input_runoff_volume_m3
        - prototype_drainage_removed_volume_m3
        - volume_conservation_error_m3
        - volume_conservation_check
...
    DrainageSummaryResponse:
      type: object
      additionalProperties: true
      required: [status]
...
    FloodDepthSummaryResponse:
      type: object
      additionalProperties: true
      required: [status]
    RoadFloodRiskResponse:
      type: object
      additionalProperties: true
      required: [status, output]