# lib/api-zod/src/generated/types/geoJsonFeatureCollection.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hyderabad Urban Flood Nowcasting System foundation API
 * OpenAPI spec version: 0.1.0
 */
import type { GeoJsonFeatureCollectionFeaturesItem } from './geoJsonFeatureCollectionFeaturesItem';

export interface GeoJsonFeatureCollection {
  type: string;
  features: GeoJsonFeatureCollectionFeaturesItem[];
  [key: string]: unknown;
 }