# lib/api-zod/src/generated/types/mapLayersResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hyderabad Urban Flood Nowcasting System foundation API
 * OpenAPI spec version: 0.1.0
 */
import type { MapLayersResponseLayers } from './mapLayersResponseLayers';
import type { MapLayersResponseLimits } from './mapLayersResponseLimits';

export interface MapLayersResponse {
  status: string;
  forecast_minutes: number;
  layers: MapLayersResponseLayers;
  limits: MapLayersResponseLimits;
}