# lib/api-zod/src/generated/types/routePoint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hyderabad Urban Flood Nowcasting System foundation API
 * OpenAPI spec version: 0.1.0
 */

export interface RoutePoint {
  lat: number;
  lon: number;
}