# lib/api-zod/src/generated/types/routeSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hyderabad Urban Flood Nowcasting System foundation API
 * OpenAPI spec version: 0.1.0
 */
import type { RoutePoint } from './routePoint';

export interface RouteSummary {
  node_sequence: string[];
  route: RoutePoint[];
  distance_m: number;
  estimated_time_min: number;
  max_flood_depth_cm: number;
  mean_flood_depth_cm: number;
  max_risk_class: number;
  safety: string;
}